backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-limiter==3.3.1
flask-caching==2.0.2
pymongo==4.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
shap==0.42.1
joblib==1.3.2
Pillow==10.0.1
reportlab==4.0.4
qrcode==7.4.2
networkx==3.1
PyJWT==2.8.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
celery==5.3.1
redis==5.0.1
pydantic==2.5.0
email-validator==2.1.0
structlog==23.2.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
mypy==1.8.0
types-requests==2.31.0.20240125